move backend logic into @openhome-core from @openhome-ui
