+++ src/ui/App.tsx
@@ -1,10 +1,10 @@
+import { BackendContext } from '@openhome-core/backend/backendContext'
+import BackendInterface from '@openhome-core/backend/backendInterface'
+import { BackendProvider } from '@openhome-core/backend/backendProvider'
+import { TauriBackend } from '@openhome-core/tauri/tauriBackend'
 import { R } from '@openhome-core/util/functional'
 import '@openhome-ui/App.css'
 import AppTabs from '@openhome-ui/AppTabs'
-import { BackendContext } from '@openhome-ui/backend/backendContext'
-import BackendInterface from '@openhome-ui/backend/backendInterface'
-import { BackendProvider } from '@openhome-ui/backend/backendProvider'
-import { TauriBackend } from '@openhome-ui/backend/tauri/tauriBackend'
 import useIsDarkMode from '@openhome-ui/hooks/darkMode'
 import useDebounce from '@openhome-ui/hooks/debounce'
 import useDisplayError from '@openhome-ui/hooks/displayError'
