+++ src/ui/hooks/isDebug.ts
@@ -1,5 +1,5 @@
+import { BackendContext } from '@openhome-core/backend/backendContext'
 import { R } from '@openhome-core/util/functional'
-import { BackendContext } from '@openhome-ui/backend/backendContext'
 import { useContext, useEffect, useState } from 'react'
 
 export default function useIsDebug(): boolean {
