+++ src/ui/pages/Settings.tsx
@@ -1,6 +1,6 @@
+import { BackendContext } from '@openhome-core/backend/backendContext'
 import { R } from '@openhome-core/util/functional'
 import { stringSorter } from '@openhome-core/util/sort'
-import { BackendContext } from '@openhome-ui/backend/backendContext'
 import ContentCard from '@openhome-ui/components/ContentCard'
 import SideTabNavigation from '@openhome-ui/components/side-tabs/SideTabNavigation'
 import { AppInfoContext, AppTheme } from '@openhome-ui/state/appInfo'
