+++ src/ui/pages/logs/LogsPage.tsx
@@ -1,8 +1,8 @@
 import { Input, Separator } from '@base-ui/react'
+import { LogEntry, LogLevel } from '@openhome-core/backend/backendInterface'
 import { OhpkmIdentifier } from '@openhome-core/pkm/Lookup'
 import { OHPKM } from '@openhome-core/pkm/OHPKM'
 import { Option } from '@openhome-core/util/functional'
-import { LogEntry, LogLevel } from '@openhome-ui/backend/backendInterface'
 import { Dialog } from '@openhome-ui/components/dialog/Dialog'
 import { ExpandIcon, FilterIcon } from '@openhome-ui/components/Icons'
 import { InfoGrid } from '@openhome-ui/components/InfoGrid'
