+++ src/ui/pages/plugins/BrowsePlugins.tsx
@@ -1,5 +1,5 @@
+import { BackendContext } from '@openhome-core/backend/backendContext'
 import { R } from '@openhome-core/util/functional'
-import { BackendContext } from '@openhome-ui/backend/backendContext'
 import DebugOnly from '@openhome-ui/components/DebugOnly'
 import { ErrorIcon } from '@openhome-ui/components/Icons'
 import useDisplayError from '@openhome-ui/hooks/displayError'
