+++ src/ui/state/convert-strategies/ConvertStrategiesProvider.tsx
@@ -1,4 +1,5 @@
-import { BackendContext } from '@openhome-ui/backend/backendContext'
+import { BackendContext } from '@openhome-core/backend/backendContext'
+import { Option } from '@openhome-core/util/functional'
 import { SyncedStateController, useSyncedState } from '@openhome-ui/state/synced-state'
 import { ConvertStrategy } from '@pkm-rs/pkg'
 import { PropsWithChildren, useContext } from 'react'
@@ -31,7 +32,10 @@ export default function ConvertStrategiesProvider({ children }: PropsWithChildre
   )
 }
 
-function stateReducer(prev: ConvertStrategies, updated: ConvertStrategies): ConvertStrategies {
+function stateReducer(
+  prev: Option<ConvertStrategies>,
+  updated: ConvertStrategies
+): ConvertStrategies {
   return { ...prev, ...updated }
 }
