+++ src/ui/state/lookups/LookupsProvider.tsx
@@ -1,11 +1,12 @@
-import { StoredLookups } from '@openhome-ui/backend/backendInterface'
+import { BackendContext } from '@openhome-core/backend/backendContext'
+import { StoredLookups } from '@openhome-core/backend/backendInterface'
+import { Option } from '@openhome-core/util/functional'
 import {
   RustStateProvider,
   SyncedStateController,
   useSyncedState,
 } from '@openhome-ui/state/synced-state'
 import { PropsWithChildren, useCallback, useContext } from 'react'
-import { BackendContext } from '../../backend/backendContext'
 import { LookupsContext } from './useLookups'
 
 function useLookupsTauri() {
@@ -23,10 +24,10 @@ export default function LookupsProvider({ children }: PropsWithChildren) {
   )
 }
 
-function stateReducer(prev: StoredLookups, updated: StoredLookups): StoredLookups {
+function stateReducer(prev: Option<StoredLookups>, updated: StoredLookups): StoredLookups {
   return {
-    gen12: { ...prev.gen12, ...updated.gen12 },
-    gen345: { ...prev.gen345, ...updated.gen345 },
+    gen12: { ...prev?.gen12, ...updated.gen12 },
+    gen345: { ...prev?.gen345, ...updated.gen345 },
   }
 }
