+++ src/ui/state/lookups/useLookups.ts
@@ -1,5 +1,5 @@
+import { StoredLookups } from '@openhome-core/backend/backendInterface'
 import { Errorable, R } from '@openhome-core/util/functional'
-import { StoredLookups } from '@openhome-ui/backend/backendInterface'
 import { createContext, useContext } from 'react'
 
 export function useLookups() {
