+++ src/ui/state/ohpkm/OhpkmStoreProvider.tsx
@@ -1,11 +1,12 @@
+import { BackendContext } from '@openhome-core/backend/backendContext'
 import { OHPKM } from '@openhome-core/pkm/OHPKM'
 import { displayIndexAdder, isBattleFormeItem } from '@openhome-core/pkm/util'
-import { BackendContext } from '@openhome-ui/backend/backendContext'
+import { StringToB64 } from '@openhome-core/tauri/tauriCommands'
+import { Option } from '@openhome-core/util/functional'
 import { SyncedStateController, useSyncedState } from '@openhome-ui/state/synced-state'
 import { PokedexUpdate } from '@openhome-ui/util/pokedex'
 import { PropsWithChildren, useCallback, useContext } from 'react'
 import { OhpkmStoreContext, OhpkmStoreData } from '.'
-import { StringToB64 } from '../../backend/tauri/tauriCommands'
 import SyncedStateProvider from '../synced-state/SyncedStateProvider'
 
 function useOhpkmStoreTauri() {
@@ -33,11 +34,11 @@ function loadOhpkmFromB64(response: StringToB64): OhpkmStoreData {
   )
 }
 
-function stateReducer(prev: OhpkmStoreData, updated: OhpkmStoreData): OhpkmStoreData {
+function stateReducer(prev: Option<OhpkmStoreData>, updated: OhpkmStoreData): OhpkmStoreData {
   return { ...prev, ...updated }
 }
 
-function useSyncedOhpkmState(): SyncedStateController<OhpkmStoreData, StringToB64> {
+function useSyncedOhpkmState(): SyncedStateController<OhpkmStoreData, OhpkmStoreData, StringToB64> {
   const backend = useContext(BackendContext)
 
   const stateGetter = backend.loadOhpkmStore
