+++ src/ui/state/plugin/PluginProvider.tsx
@@ -1,5 +1,5 @@
+import { BackendContext } from '@openhome-core/backend/backendContext'
 import { partitionResults, R } from '@openhome-core/util/functional'
-import { BackendContext } from '@openhome-ui/backend/backendContext'
 import useDisplayError from '@openhome-ui/hooks/displayError'
 import { GITHUB_REPO, LOCAL_REPO } from '@openhome-ui/pages/plugins/BrowsePlugins'
 import { CURRENT_PLUGIN_API_VERSION } from '@openhome-ui/pages/plugins/Plugins'
