+++ src/ui/state/synced-state/useSyncedState.ts
@@ -1,5 +1,5 @@
+import { BackendContext } from '@openhome-core/backend/backendContext'
 import { Errorable, Option, R } from '@openhome-core/util/functional'
-import { BackendContext } from '@openhome-ui/backend/backendContext'
 import { useCallback, useContext, useEffect, useState } from 'react'
 
 type StateConverter<State, RustState> = [State] extends [RustState]
