+++ src/ui/util/plugin.ts
@@ -1,5 +1,5 @@
+import { ImageResponse } from '@openhome-core/backend/backendInterface'
 import { NationalDex } from '@openhome-core/resources/consts/NationalDex'
-import { ImageResponse } from '@openhome-ui/backend/backendInterface'
 import { OpenHomePlugin } from '@openhome-ui/state/plugin/reducer'
 import { MetadataSummaryLookup, SpeciesLookup } from '@pkm-rs/pkg'
